+++ tests/argv.hpp
@@ -16,12 +16,12 @@ class Argv {
  public:
   explicit Argv(std::initializer_list<const char*> args);
   char** argv() const;
-  size_t argc() const;
+  int argc() const;
 
  private:
   std::vector<std::unique_ptr<char[]>> m_args;
   std::unique_ptr<char*[]> m_argv;
-  size_t m_argc;
+  int m_argc;
 };
 
 #endif  // ZEPHIR_TEST_ARGV_HPP_
